requirements.txt
pathlib>=1.0.0
raven
django-picker
django-sass-processor
django-compressor
libsass
django~=2.1.8
celery==3.1.25
django-celery==3.2.2
redis==2.10.6
